Better usb interface and added macro keys, and a command channel
The usb enumeration required all interface to be enumerated before
freeing the stack to do keyboard interactions. The trouble with this
approach is that:
- linux console does not require a mouse, and does not configure that
  interface
- windows requires a driver for the serial port, and does not
  configure that interface

Now we enumerate for all of 2 seconds, and if some interfaces are
configured, we say -- good to go!

Adding macro keys requires:
- proper handling of a host being ready for the next key; we now
  attach to the interrupt out of the host, so know when it is ok to send
  more.
- translation from ascii to usb key codes -- so anything not in 7 bit
  clean is not supported.

Adding a command channel -- press '?' to get an idea of commands that
are accepted.