Improve USB suspend handling
I misunderstood usb suspend semantics; it is used to indicate that the
host has had no communication for us for more than 3ms and that we are
ok to go to a lower power state.

This does not change the fact that we might have something to say to
the host. So even in lower power state, we should keep monitoring our
keys/mouse and issue a "remote wakeup" when such an event occurs.