User flash actions simplified
User flashing of configuration is now managed in two functions:
flash_read_config and flash_write_config.

The previous interface was too fragile and took a lot more code. This
new approach captures the actual structures in flash in terms of
structs. This allows the compiler to check our bounds during
compilation.